LPS-179353 Extract ScreenNavigationEntry methods from CommerceInventoryWarehouseQualifiersScreenNavigationCategory to CommerceInventoryWarehouseQualifiersScreenNavigationEntry, extends CommerceInventoryWarehouseQualifiersScreenNavigationCategory in it and return getCategoryKey in getEntryKey method
