+++ modules/apps/commerce/commerce-warehouse-web/src/main/java/com/liferay/commerce/warehouse/web/internal/frontend/taglib/servlet/taglib/CommerceInventoryWarehouseQualifiersScreenNavigationCategory.java
@@ -15,36 +15,17 @@
 package com.liferay.commerce.warehouse.web.internal.frontend.taglib.servlet.taglib;
 
 import com.liferay.commerce.country.CommerceCountryManager;
-import com.liferay.commerce.inventory.model.CommerceInventoryWarehouse;
 import com.liferay.commerce.inventory.service.CommerceInventoryWarehouseRelService;
-import com.liferay.commerce.inventory.service.CommerceInventoryWarehouseService;
-import com.liferay.commerce.product.service.CommerceChannelRelService;
 import com.liferay.commerce.product.service.CommerceChannelService;
 import com.liferay.commerce.warehouse.web.internal.constants.CommerceInventoryWarehouseScreenNavigationConstants;
-import com.liferay.commerce.warehouse.web.internal.display.context.CommerceInventoryWarehouseQualifiersDisplayContext;
 import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationCategory;
-import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationEntry;
-import com.liferay.frontend.taglib.servlet.taglib.util.JSPRenderer;
 import com.liferay.portal.kernel.language.Language;
-import com.liferay.portal.kernel.log.Log;
-import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.kernel.model.User;
-import com.liferay.portal.kernel.security.permission.ActionKeys;
-import com.liferay.portal.kernel.security.permission.PermissionThreadLocal;
-import com.liferay.portal.kernel.security.permission.resource.ModelResourcePermission;
 import com.liferay.portal.kernel.service.CountryService;
 import com.liferay.portal.kernel.service.RegionService;
-import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
-import com.liferay.portal.kernel.util.WebKeys;
-
-import java.io.IOException;
 
 import java.util.Locale;
 import java.util.ResourceBundle;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -53,24 +34,14 @@ import org.osgi.service.component.annotations.Reference;
  * @author Andrea Sbarra
  */
 @Component(
-	property = {
-		"screen.navigation.category.order:Integer=20",
-		"screen.navigation.entry.order:Integer=10"
-	},
-	service = {ScreenNavigationCategory.class, ScreenNavigationEntry.class}
+	property = "screen.navigation.category.order:Integer=20",
+	service = ScreenNavigationCategory.class
 )
 public class CommerceInventoryWarehouseQualifiersScreenNavigationCategory
-	implements ScreenNavigationCategory,
-			   ScreenNavigationEntry<CommerceInventoryWarehouse> {
+	implements ScreenNavigationCategory {
 
 	@Override
 	public String getCategoryKey() {
-		return CommerceInventoryWarehouseScreenNavigationConstants.
-			CATEGORY_KEY_QUALIFIERS;
-	}
-
-	@Override
-	public String getEntryKey() {
 		return CommerceInventoryWarehouseScreenNavigationConstants.
 			CATEGORY_KEY_QUALIFIERS;
 	}
@@ -80,7 +51,7 @@ public class CommerceInventoryWarehouseQualifiersScreenNavigationCategory
 		ResourceBundle resourceBundle = ResourceBundleUtil.getBundle(
 			"content.Language", locale, getClass());
 
-		return _language.get(resourceBundle, "eligibility");
+		return language.get(resourceBundle, "eligibility");
 	}
 
 	@Override
@@ -89,90 +60,23 @@ public class CommerceInventoryWarehouseQualifiersScreenNavigationCategory
 			SCREEN_NAVIGATION_KEY_COMMERCE_INVENTORY_WAREHOUSE_GENERAL;
 	}
 
-	@Override
-	public boolean isVisible(
-		User user, CommerceInventoryWarehouse commerceInventoryWarehouse) {
-
-		if (commerceInventoryWarehouse == null) {
-			return false;
-		}
-
-		boolean hasPermission = false;
-
-		try {
-			hasPermission =
-				_commerceInventoryWarehouseModelResourcePermission.contains(
-					PermissionThreadLocal.getPermissionChecker(),
-					commerceInventoryWarehouse.
-						getCommerceInventoryWarehouseId(),
-					ActionKeys.UPDATE);
-		}
-		catch (Exception exception) {
-			if (_log.isDebugEnabled()) {
-				_log.debug(exception);
-			}
-		}
-
-		return hasPermission;
-	}
-
-	@Override
-	public void render(
-			HttpServletRequest httpServletRequest,
-			HttpServletResponse httpServletResponse)
-		throws IOException {
-
-		httpServletRequest.setAttribute(
-			WebKeys.PORTLET_DISPLAY_CONTEXT,
-			new CommerceInventoryWarehouseQualifiersDisplayContext(
-				_commerceChannelRelService, _commerceInventoryWarehouseService,
-				httpServletRequest, _portal,
-				_commerceInventoryWarehouseModelResourcePermission));
-
-		_jspRenderer.renderJSP(
-			httpServletRequest, httpServletResponse,
-			"/commerce_inventory_warehouse/qualifiers.jsp");
-	}
-
-	private static final Log _log = LogFactoryUtil.getLog(
-		CommerceInventoryWarehouseQualifiersScreenNavigationCategory.class);
+	@Reference
+	protected CommerceChannelService commerceChannelService;
 
 	@Reference
-	private CommerceChannelRelService _commerceChannelRelService;
+	protected CommerceCountryManager commerceCountryManager;
 
 	@Reference
-	private CommerceChannelService _commerceChannelService;
+	protected CommerceInventoryWarehouseRelService
+		commerceInventoryWarehouseRelService;
 
 	@Reference
-	private CommerceCountryManager _commerceCountryManager;
-
-	@Reference(
-		target = "(model.class.name=com.liferay.commerce.inventory.model.CommerceInventoryWarehouse)"
-	)
-	private ModelResourcePermission<CommerceInventoryWarehouse>
-		_commerceInventoryWarehouseModelResourcePermission;
+	protected CountryService countryService;
 
 	@Reference
-	private CommerceInventoryWarehouseRelService
-		_commerceInventoryWarehouseRelService;
+	protected Language language;
 
 	@Reference
-	private CommerceInventoryWarehouseService
-		_commerceInventoryWarehouseService;
-
-	@Reference
-	private CountryService _countryService;
-
-	@Reference
-	private JSPRenderer _jspRenderer;
-
-	@Reference
-	private Language _language;
-
-	@Reference
-	private Portal _portal;
-
-	@Reference
-	private RegionService _regionService;
+	protected RegionService regionService;
 
 }
